miradorstack-playground-main/requirements.txt
requests
opentelemetry-distro
opentelemetry-exporter-otlp-proto-http